refactor: mv streamToArray to util.js and rename
